+++ getMasterThemes.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-git clone https://github.com/Unthrottled/doki-master-theme.git masterThemes
+git clone https://github.com/doki-theme/doki-master-theme.git masterThemes